refactor(http): split request option types
